# docs/swagger.yaml
basePath: /
definitions:
  dto.Comment:
    properties:
      message:
        type: string
      photo_id:
        example: 1
        type: integer
    required:
    - message
    - photo_id
    type: object
  dto.CommentMessage:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  dto.Photo:
    properties:
      caption:
        type: string
      photo_url:
        example: https://subdomain.domain.dom.ge/path?arg=1
        type: string
      title:
        type: string
    required:
    - photo_url
    - title
    type: object
...
  dto.UserUpdate:
    properties:
      email:
        example: name@org.dom.ge
        type: string
      username:
        type: string
    required:
    - email
    - username
    type: object
  models.Comment:
    properties:
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      message:
        type: string
      photo_id:
        example: 1
        type: integer
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  models.Photo:
    properties:
      caption:
        type: string
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://subdomain.domain.dom.ge/path?arg=1
        type: string
      title:
        type: string
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
  responses.CreateComment:
    properties:
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      message:
        type: string
      photo_id:
        example: 1
        type: integer
      user_id:
        example: 1
        type: integer
    type: object
  responses.CreatePhoto:
    properties:
      caption:
        type: string
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://subdomain.domain.dom.ge/path?arg=1
        type: string
      title:
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  responses.ErrorMessage:
    properties:
      error_message:
        type: string
    type: object
...
  responses.GetComment:
    properties:
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      message:
        type: string
      photo:
        $ref: '#/definitions/models.Photo'
      photo_id:
        example: 1
        type: integer
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user:
        $ref: '#/definitions/responses.UserComment'
      user_id:
        example: 1
        type: integer
    type: object
  responses.GetPhoto:
    properties:
      caption:
        type: string
      created_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://subdomain.domain.dom.ge/path?arg=1
        type: string
      title:
        type: string
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user:
        $ref: '#/definitions/dto.UserUpdate'
      user_id:
        example: 1
        type: integer
    type: object
...
  responses.Message:
    properties:
      message:
        type: string
    type: object
  responses.UpdatePhoto:
    properties:
      caption:
        type: string
      id:
        example: 1
        type: integer
      photo_url:
        example: https://subdomain.domain.dom.ge/path?arg=1
        type: string
      title:
        type: string
      updated_at:
        example: "2019-11-09T21:21:46+00:00"
        type: string
      user_id:
        example: 1
        type: integer
    type: object
...
  responses.UserComment:
    properties:
      email:
        example: name@org.dom.ge
        type: string
      id:
        example: 1
        type: integer
      username:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: bastianarf@google.com
    name: Bastian Arfianto
  description: API server for MyGram social media in "Scalable Webservice with Golang"
    course from Hacktiv8 × Kominfo.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MyGram API
  version: "1.0"
paths:
  /comments:
    get:
      consumes:
      - application/json
      description: Get comments.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.GetComment'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      description: Create a Comment associated with the logged in user.
      parameters:
      - description: JSON of the comment to be made. Caption is not mandatory.
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dto.Comment'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CreateComment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a Comment
      tags:
      - comments
  /comments/{commentId}:
    delete:
      consumes:
      - application/json
      description: Delete a comment associated with logged in user.
      parameters:
      - description: ID number of the comment to be deleted
        in: path
        name: commentId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
    put:
      consumes:
      - application/json
      description: Update a comment associated with logged in user.
      parameters:
      - description: ID number of the comment
        in: path
        name: commentId
        required: true
        type: integer
      - description: New JSON of the comment.
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/dto.CommentMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Comment'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update a comment
      tags:
      - comments
  /photos:
    get:
      consumes:
      - application/json
      description: Get photos.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/responses.GetPhoto'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Get photos
      tags:
      - photos
    post:
      consumes:
      - application/json
      description: Create a Photo associated with the logged in user identified by
        bearer token.
      parameters:
      - description: JSON of the photo to be made. Caption is not mandatory.
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/dto.Photo'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.CreatePhoto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Create a Photo
      tags:
      - photos
  /photos/{photoId}:
    delete:
      consumes:
      - application/json
      description: Delete a photo associated with logged in user.
      parameters:
      - description: ID number of the photo to be deleted
        in: path
        name: photoId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Delete a photo
      tags:
      - photos
    put:
      consumes:
      - application/json
      description: Update a photo associated with logged in user.
      parameters:
      - description: ID number of the photo
        in: path
        name: photoId
        required: true
        type: integer
      - description: New JSON of the photo.
        in: body
        name: photo
        required: true
        schema:
          $ref: '#/definitions/dto.Photo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.UpdatePhoto'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorMessage'
        "500":
          description: Internal Server Error
      security:
      - BearerAuth: []
      summary: Update a photo
      tags:
      - photos
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"